+++ R/new_coin.R
@@ -105,6 +105,10 @@
 #' @param level_names Optional character vector of names of levels. Must have length equal to the number of
 #' levels in the hierarchy (`max(iMeta$Level, na.rm = TRUE)`).
 #' @param quietly If `TRUE`, suppresses all messages
+#' @param retain_all_uCodes_on_split Logical: if panel data is input and split to a purse using `split_to`, this controls how
+#' units with no data at certain time points are handled. If set `FALSE`, then unit at time t with no data in any indicators
+#' will be removed completely from the coin for that time point. If `TRUE`, all units will be included in every time point. The latter option
+#' may be useful if you impute over time.
 #'
 #' @examples
 #' # build a coin using example data frames
@@ -128,7 +132,7 @@
 #'
 #' @export
 new_coin <- function(iData, iMeta, exclude = NULL, split_to = NULL,
-                     level_names = NULL, quietly = FALSE){
+                     level_names = NULL, retain_all_uCodes_on_split = FALSE, quietly = FALSE){
 
   # WRITE TO LOG ------------------------------------------------------------
 
@@ -244,7 +248,7 @@ new_coin <- function(iData, iMeta, exclude = NULL, split_to = NULL,
     }
 
     # now split
-    iData_list <- split_iData(iData, split_to = split_to)
+    iData_list <- split_iData(iData, split_to = split_to, retain_all_uCodes = retain_all_uCodes_on_split)
     # check
     suppressMessages(lapply(iData_list, check_iData))
   } else {
@@ -368,6 +372,11 @@ check_iData <- function(iData, quietly = FALSE){
   # check type
   if(!is.character(iData$uCode)){
     stop("uCode is required to be a character vector.")
+  }
+  # should not contain spaces
+  spaces <- grepl(" ", iData$uCode)
+  if(any(spaces)){
+    stop("uCode contains entries with spaces - this is not allowed.")
   }
 
   # SPECIAL COLS ------------------------------------------------------------
@@ -707,7 +716,7 @@ check_iMeta <- function(iMeta, quietly = FALSE){
 # vector.
 #
 # @return List of `iData` data frames
-split_iData <- function(iData, split_to){
+split_iData <- function(iData, split_to, retain_all_uCodes = FALSE){
 
   # this function is only called from new_coin(), so if we are here, then the iData should be valid,
   # and there should be more than one unique entry in iData$Year.
@@ -721,7 +730,29 @@ split_iData <- function(iData, split_to){
   }
 
   # return list of dfs
-  split(iData, iData$Time)
+  l_iData <- split(iData, iData$Time)
+
+  if(retain_all_uCodes){
+    # full set of ucodes
+    uCodes <- unique(iData$uCode)
+    # filler df of NAs to use to fill gaps
+    iDataNA <- iData[match(uCodes, iData$uCode), ]
+    iDataNA[names(iDataNA) %nin% c("uCode", "uName", "Time")] <- NA
+
+
+    # check each time point one at a time
+    l_iData <- lapply(l_iData, function(X){
+      missing_ucodes <- uCodes[uCodes %nin% X$uCode]
+      # exit if no uCodes missing
+      if(length(missing_ucodes) == 0) return(X)
+      # else add NA rows from the df above
+      rows_to_add <- iDataNA[iDataNA$uCode %in% missing_ucodes, ]
+      rows_to_add$Time <- unique(X$Time)
+      rbind(X, rows_to_add)
+    })
+  }
+
+  l_iData
 
 }
 
@@ -749,14 +780,14 @@ get_lineage <- function(iMeta, level_names = NULL){
   if(maxlev == 1){
     wideS <- wideS["iCode"]
     warning("Only one level is defined in iMeta. This is not normally expected in a composite indicator, and some functions may not work as expected.", call. = FALSE)
-  } else {
+  } else if (maxlev > 2) {
     # successively add columns by looking up parent codes of last col
     for(ii in 2:(maxlev-1)){
       wideS <- cbind(wideS,
                      longS$Parent[match(wideS[[ii]], longS$iCode)])
     }
   }
-
+  # NOTE: if maxlev == 2, nothing needs to be done
 
   # rename columns
   if(is.null(level_names)){